+++ lib/src/object_extension.dart
@@ -5,7 +5,7 @@ extension ObjectExtension on Object {
   /// Determines if this [Object] is empty.
   bool get isEmpty {
     if (this is String) {
-      return toString().trim().isEmpty;
+      return toString().isEmpty;
     }
     if (this is Iterable) {
       final Iterable<dynamic> value = this as Iterable<dynamic>;
@@ -17,10 +17,29 @@ extension ObjectExtension on Object {
     }
     return false;
   }
+
+  bool get isBlank {
+    if (this is String) {
+      return toString().isBlank;
+    }
+    if (this is Iterable) {
+      final Iterable<Object?> value = this as Iterable<Object?>;
+      return value.every((Object? e) => e.isBlank);
+    }
+    if (this is Map) {
+      final Map<Object?, Object?> value = this as Map<Object?, Object?>;
+      return value.entries.every(
+        (MapEntry<Object?, Object?> e) => e.value.isBlank,
+      );
+    }
+    return false;
+  }
 }
 
 /// Nullable [Object] Extension.
 extension NullableObjectExtension on Object? {
   /// Determines if this [Object] is empty.
-  bool get isEmpty => this != null && this!.isEmpty;
+  bool get isEmpty => this == null || this!.isBlank;
+
+  bool get isBlank => this == null || this!.isBlank;
 }